86 Expo, switch servo player 2 buttons, ballLaunch
This commit is to add the files that were fixed on Wednesday, two days
before the competition - TeleOp was not working with the joy sticks of
moving straight with the commit merge before. we resolved this by taking
out the Expo.

The ball collector buttons were backwards for the ball launcher - this had
been fixed as well.

Another autonmous method have been added for a sleep between the ball
launches.