Adjust spin wheel to 0.7, +strafing speed in Tank
Duchess has all 4 legs again. The SW team ran through the test programs
designed to test each encoder motor by turning right/left, strafing from
side to side, moving fowards and backwards, and run the spinner wheel
motor for five seconds. each test passed with the exemption of strafing
was a bit off - we think it's due to the weight of the robot on the front
left wheel which causes that specific wheel to not move and it gets
dragged behind.

After the tests passed, the team adjusted the spinner wheel to be 0.7. we
can see the values are being passed through. The driving team is
practicing with this speed for launching balls.

DRiving team also examined that the strafing in Tank TeleOp mode is slow.
SW attempted to change this value to a higher strafing power - driving
testing still needs to be observed.

Finally the attempt to add in the Beacon servo arm (s1) to gamepad 2 still
needs to be fixed. A couple of hardware configurations in the hardwareMap
for SW was played around to see if it's the gamepad wasn't reading the
servo class correctly - testing needs to be done for this.